Generate counter-models without full traces
We found that traces can take a very long time to generate
(~15 minutes for some of our use cases) due to the `match_annotation`
mechanism. We didn't deeply investigate the reasons for the bottleneck,
but `match_annotation` is implemented in a recursive style, and it might
be the case that it does a lot of copying in memory.

Added two CLI options:
- `model` (false by default)
- `shrink` (true by default; only has an effect if model=true)

`model` tells Ivy to only generate and pretty-print the model,
not the full trace. `shrink` controls whether the returned model
is minimised or not.