de Bruijn looks OK; trouble with modifies again
It seems like the the issue with de Bruijn indices was trivial
to fix.

I am now having issues with mypyvy's modifies again. Since we
now simplify formulas, we cannot rely on Ivy's 'modified' to generate
the mypyvy one (even with the previous caveats).

I had removed the code that distinguishes between "actually modified"
and "only perceived by mypyvy to be modified" fields, but we cannot
actually do this. In the below, t__this is considered modified, so
it's havoc'ed if we don't explicitly set it:

```
new(env_historical_auth_required(O, t__this, _approve)
```

I'll need to reimplement this logic, which is now a bit trickier to
do because of the simplification.