WIP: remove Skolems
This is enough to simplify `mint` for the LP contract (modulo the
FIXME in `is_temp_equality`), but fails on `deposit` and other
transitions because it tries to unfold:

```
[_macro_139] __fml:depositor = pool._depositor_addr
unfolding macro __fml:depositor = pool._depositor_addr...
```

This should NOT happen because `depositor` is a formal parameter
of the `deposit` transition.

This also seems to originate from `if depositor = _depositor_addr` in the
GHOST portion of the spec, so really shouldn't be identified as a macro even
if `depositor` wasn't a parameter.

Will investigate.